% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate.patterns.functions.R
\name{extraction_Lmm_from_rank}
\alias{extraction_Lmm_from_rank}
\title{This function will convert the rank test in a table with p.value, pattern of the spp (random, aggregate or repuslion)}
\usage{
extraction_Lmm_from_rank(rank_result)
}
\arguments{
\item{rank_result}{the output of the rank result for dbh}
}
\description{
This function will convert the rank test in a table with p.value, pattern of the spp (random, aggregate or repuslion)
}
\examples{
extraction_Lmm_from_rank (rank_test_replicated_500)
}
\keyword{Lmm}

