% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate.patterns.functions.R
\name{replicate_patterns}
\alias{replicate_patterns}
\title{This function will compute replicated patterns from envelope data}
\usage{
replicate_patterns(Status, method, mark, pattern_to_murge)
}
\arguments{
\item{Status}{"living" or "dead"}

\item{method}{It should be 'normal' or 'size500' depenging on the method to construct replicated patterns. Need to compute before  dbh_pattern_to_murge. Defaults to "normal"}

\item{mark}{NA, "DBH" or "species"}

\item{pattern_to_murge}{list with Country_ID and stand pattern}
}
\value{
list by countrie of replicated spp for each stand
}
\description{
This function will compute replicated patterns from envelope data
}
\examples{
eplicate_patterns(Status="living",mark=NA,method="size500",dbh_pattern_to_murge)

}
\keyword{replicated_pattern}

