% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate.patterns.functions.R
\name{replicate_patterns_to_rank_test}
\alias{replicate_patterns_to_rank_test}
\title{This function will compute the rank test from replicated patterns}
\usage{
replicate_patterns_to_rank_test(Status, replicate_patterns, mark, graph)
}
\arguments{
\item{Status}{"living" or "dead"}

\item{replicate_patterns}{the output of the replicate_patterns function}

\item{mark}{description of the mark 'DBH', or NA is none.}

\item{graph}{TRUE or FALSE. Defaults to FALSE}
}
\description{
This function will compute the rank test from replicated patterns
}
\examples{
rank_test_replicated_dbh_500 <- replicate_patterns_to_rank_test(replicate_patterns_dbh_size500,"dbh",graph=TRUE)
}
\keyword{replicate_patterns}

